+++ apps/cli/lib/src/cli/cli_runtime.dart
@@ -1,0 +1,17 @@
+import 'package:celest_cli/src/context.dart';
+
+enum CliRuntime {
+  pubGlobal,
+  aot,
+  local;
+
+  static CliRuntime get current {
+    if (fileSystem.path.fromUri(platform.script).endsWith('.snapshot')) {
+      return pubGlobal;
+    } else if (platform.executable.contains('dart')) {
+      return local;
+    } else {
+      return aot;
+    }
+  }
+}