+++ apps/cli/lib/src/frontend/resident_compiler.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:celest_cli/src/context.dart';
 import 'package:celest_cli/src/sdk/dart_sdk.dart';
-import 'package:celest_cli/src/utils/cli.dart';
+import 'package:celest_cli/src/utils/process.dart';
 import 'package:file/file.dart';
 import 'package:logging/logging.dart';